backend/requirements.txt
python-dotenv
"langchain[docarray]"
langchain
langchain-openai
pypdf
beautifulsoup4
requests